fix: change name, change global style
