+++ src/app/page.tsx
@@ -1,3 +1,11 @@
+import { RadioGroup, RadioItem } from '@/shared/components/Radio-group'
+
 export default function Home() {
-   return <div className="text-xxl font-bold">Test</div>
+   return (
+      <RadioGroup>
+         <RadioItem value="1" label="Первый вариант" />
+         <RadioItem value="2" label="Второй вариант" />
+         <RadioItem value="3" label="Третий вариант" />
+      </RadioGroup>
+   )
 }
