+++ src/shared/components/Radio-group/index.ts
@@ -1,0 +1,2 @@
+export { RadioGroup } from './Radio-group'
+export { RadioItem } from './RadioItem'
