--- src/styles/boilerplate.css
@@ -1,5 +1,0 @@
-html {
-    line-height: var(--line-height-m);
-    color: var(--color-light-100);
-    background-color: var(--color-dark-700);
-}
