+++ src/styles/index.css
@@ -1,3 +1,3 @@
 @import './colors.css';
 @import './typography.css';
-@import './boilerplate.css';
+/* @import './boilerplate.css'; */
